Support pre-match async filters
