+++ resteasy-jaxrs/src/main/java/org/jboss/resteasy/core/interception/jaxrs/PreMatchContainerRequestContext.java
@@ -1,9 +1,12 @@
 package org.jboss.resteasy.core.interception.jaxrs;
 
+import org.jboss.resteasy.specimpl.BuiltResponse;
+import org.jboss.resteasy.spi.ApplicationException;
 import org.jboss.resteasy.spi.HttpRequest;
 import org.jboss.resteasy.spi.ResteasyProviderFactory;
 
 import javax.ws.rs.container.ContainerRequestContext;
+import javax.ws.rs.container.ContainerRequestFilter;
 import javax.ws.rs.core.Cookie;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.MultivaluedMap;
@@ -11,6 +14,8 @@ import javax.ws.rs.core.Request;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.SecurityContext;
 import javax.ws.rs.core.UriInfo;
+
+import java.io.IOException;
 import java.io.InputStream;
 import java.net.URI;
 import java.util.ArrayList;
@@ -29,10 +34,20 @@ public class PreMatchContainerRequestContext implements ContainerRequestContext
 {
    protected final HttpRequest httpRequest;
    protected Response response;
+   private ContainerRequestFilter[] requestFilters;
+   private int currentFilter;
+   private boolean suspended;
+   private boolean filterReturnIsMeaningful = true;
+   private Runnable continuation;
+   private Map<Class<?>, Object> contextDataMap;
 
-   public PreMatchContainerRequestContext(HttpRequest request)
+   public PreMatchContainerRequestContext(HttpRequest request, 
+         ContainerRequestFilter[] requestFilters, Runnable continuation)
    {
       this.httpRequest = request;
+      this.requestFilters = requestFilters;
+      this.continuation = continuation;
+      contextDataMap = ResteasyProviderFactory.getContextDataMap();
    }
 
    public HttpRequest getHttpRequest()
@@ -190,14 +205,83 @@ public class PreMatchContainerRequestContext implements ContainerRequestContext
    }
 
    @Override
-   public void abortWith(Response response)
-   {
-      this.response = response;
-   }
-
-   @Override
    public String getHeaderString(String name)
    {
       return httpRequest.getHttpHeaders().getHeaderString(name);
+   }
+
+   public synchronized void suspend() {
+      if(continuation == null)
+         throw new RuntimeException("Suspend not supported yet");
+      suspended = true;
+   }
+
+   @Override
+   public synchronized void abortWith(Response response)
+   {
+      ResteasyProviderFactory.pushContextDataMap(contextDataMap);
+      if(suspended)
+      {
+         httpRequest.getAsyncContext().getAsyncResponse().resume(response);
+      }
+      else
+      {
+         this.response = response;
+      }
+   }
+   
+   public synchronized void resume() {
+      if(!suspended)
+         throw new RuntimeException("Cannot resume: not suspended");
+      ResteasyProviderFactory.pushContextDataMap(contextDataMap);
+      // just go on
+      filter();
+   }
+   
+   public synchronized BuiltResponse filter()
+   {
+      // FIXME: check what happens if the filter suspends and resumes/abort within the same call (same thread)
+      while(currentFilter < requestFilters.length)
+      {
+         ContainerRequestFilter filter = requestFilters[currentFilter++];
+         try
+         {
+            suspended = false;
+            filter.filter(this);
+         }
+         catch (IOException e)
+         {
+            throw new ApplicationException(e);
+         }
+         if(suspended) {
+            if(!httpRequest.getAsyncContext().isSuspended())
+               httpRequest.getAsyncContext().suspend();
+            // ignore any abort request until we are resumed
+            filterReturnIsMeaningful = false;
+            response = null;
+            return null;
+         }
+         BuiltResponse serverResponse = (BuiltResponse)getResponseAbortedWith();
+         if (serverResponse != null)
+         {
+            // handle the case where we've been suspended by a previous filter
+            if(filterReturnIsMeaningful)
+               return serverResponse;
+            else
+               httpRequest.getAsyncContext().getAsyncResponse().resume(serverResponse);
+         }
+      }
+      // here it means we reached the last filter
+      // if we've never been suspended, the caller is valid and let it go on doing the request
+      if(filterReturnIsMeaningful)
+         return null;
+      // if we've been suspended then the caller is a filter and have to invoke our continuation
+      continuation.run();
+      return null;
+   }
+
+   public boolean isSuspended()
+   {
+      return !filterReturnIsMeaningful;
    }
 }
