+++ resteasy-legacy/src/main/java/org/jboss/resteasy/core/interception/PreMatchContainerRequestContext.java
@@ -13,6 +13,6 @@ public class PreMatchContainerRequestContext extends org.jboss.resteasy.core.int
 
    public PreMatchContainerRequestContext(HttpRequest request)
    {
-      super(request);
+      super(request, null, null);
    }
 }
